Updated Reclaim Protocol

+++ src/peer/PeerInterface.java
@@ -7,6 +7,6 @@ public interface PeerInterface extends Remote {
     void backup(String path,int replicationDeg) throws RemoteException;
     void delete(String path) throws RemoteException;
     void restore(String path) throws RemoteException;
-    void reclaim(float space) throws RemoteException;
+    void reclaim(long space) throws RemoteException;
     String state() throws  RemoteException;
 }
